+++ src/main/java/com/headstrongpro/desktop/modelCollections/DBPayment.java
@@ -2,10 +2,13 @@ package com.headstrongpro.desktop.modelCollections;
 
 import com.headstrongpro.desktop.core.connection.DBConnect;
 import com.headstrongpro.desktop.core.exception.ConnectionException;
+import com.headstrongpro.desktop.core.exception.DatabaseOutOfSyncException;
 import com.headstrongpro.desktop.core.exception.ModelSyncException;
 import com.headstrongpro.desktop.model.Payment;
 import com.headstrongpro.desktop.model.Subscription;
+import com.headstrongpro.desktop.modelCollections.util.ActionType;
 import com.headstrongpro.desktop.modelCollections.util.IDataAccessObject;
+import com.headstrongpro.desktop.modelCollections.util.Synchronizable;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -13,7 +16,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DBPayment implements IDataAccessObject<Payment> {
+public class DBPayment extends Synchronizable implements IDataAccessObject<Payment> {
 
     private DBConnect dbConnect;
 
@@ -70,6 +73,7 @@ public class DBPayment implements IDataAccessObject<Payment> {
     public Payment create(Payment object) throws ModelSyncException {
         try {
             dbConnect = new DBConnect();
+            //language=TSQL
             String createPaymentQuery = "INSERT INTO payments (subscription_id, value, timestamp, due_date, paid) VALUES (?, ?, ?, ?, ?);";
             PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(createPaymentQuery, PreparedStatement.RETURN_GENERATED_KEYS);
             preparedStatement.setInt(1, object.getSubscription().getId());
@@ -81,6 +85,7 @@ public class DBPayment implements IDataAccessObject<Payment> {
             try (ResultSet generatedKeys = preparedStatement.getGeneratedKeys()) {
                 if (generatedKeys.next()) {
                     object.setId(generatedKeys.getInt(1));
+                    logChange("payments", object.getId(), ActionType.CREATE);
                 } else {
                     throw new ModelSyncException("Creating payment failed. No ID retrieved!");
                 }
@@ -92,32 +97,44 @@ public class DBPayment implements IDataAccessObject<Payment> {
     }
 
     @Override
-    public void update(Payment object) throws ModelSyncException {
-        try {
-            dbConnect = new DBConnect();
-            String query = "UPDATE payments SET subscription_id=?,value=?,timestamp=?,due_date=?,paid=? WHERE id=?;";
-            PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
-            preparedStatement.setInt(1, object.getSubscription().getId());
-            preparedStatement.setDouble(2, object.getValue());
-            preparedStatement.setDate(3, object.getTimestamp());
-            preparedStatement.setDate(4, object.getDueDate());
-            preparedStatement.setBoolean(5, object.isPaid());
-            dbConnect.uploadSafe(preparedStatement);
-        } catch (ConnectionException | SQLException e) {
-            throw new ModelSyncException("WARNING! Could not update payment of ID: " + object.getId() + " !", e);
+    public void update(Payment object) throws ModelSyncException, DatabaseOutOfSyncException {
+        if (verifyIntegrity(object.getId())){
+            try {
+                dbConnect = new DBConnect();
+                //language=TSQL
+                String query = "UPDATE payments SET subscription_id=?,value=?,timestamp=?,due_date=?,paid=? WHERE id=?;";
+                PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
+                preparedStatement.setInt(1, object.getSubscription().getId());
+                preparedStatement.setDouble(2, object.getValue());
+                preparedStatement.setDate(3, object.getTimestamp());
+                preparedStatement.setDate(4, object.getDueDate());
+                preparedStatement.setBoolean(5, object.isPaid());
+                dbConnect.uploadSafe(preparedStatement);
+                logChange("payments", object.getId(), ActionType.UPDATE);
+            } catch (ConnectionException | SQLException e) {
+                throw new ModelSyncException("WARNING! Could not update payment of ID: " + object.getId() + " !", e);
+            }
+        } else {
+            throw new DatabaseOutOfSyncException();
         }
     }
 
     @Override
-    public void delete(Payment object) throws ModelSyncException {
-        try {
-            dbConnect = new DBConnect();
-            String query = "DELETE FROM payments WHERE id=?;";
-            PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
-            preparedStatement.setInt(1, object.getId());
-            preparedStatement.execute();
-        } catch (ConnectionException | SQLException e) {
-            throw new ModelSyncException("WARNING! Could not delete payment of ID: " + object.getId() + " !", e);
+    public void delete(Payment object) throws ModelSyncException, DatabaseOutOfSyncException {
+        if (verifyIntegrity(object.getId())){
+            try {
+                dbConnect = new DBConnect();
+                //language=TSQL
+                String query = "DELETE FROM payments WHERE id=?;";
+                PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
+                preparedStatement.setInt(1, object.getId());
+                preparedStatement.execute();
+                logChange("payments", object.getId(), ActionType.DELETE);
+            } catch (ConnectionException | SQLException e) {
+                throw new ModelSyncException("WARNING! Could not delete payment of ID: " + object.getId() + " !", e);
+            }
+        } else {
+            throw new DatabaseOutOfSyncException();
         }
     }
 
@@ -143,5 +160,10 @@ public class DBPayment implements IDataAccessObject<Payment> {
             throw new ModelSyncException("Could not load subscriptions.", e);
         }
         return payments;
+    }
+
+    @Override
+    protected boolean verifyIntegrity(int itemID) throws ModelSyncException {
+        return true; //TODO: to be implemented
     }
 }
