+++ src/main/java/com/headstrongpro/desktop/modelCollections/DBSubscriptions.java
@@ -2,11 +2,14 @@ package com.headstrongpro.desktop.modelCollections;
 
 import com.headstrongpro.desktop.core.connection.DBConnect;
 import com.headstrongpro.desktop.core.exception.ConnectionException;
+import com.headstrongpro.desktop.core.exception.DatabaseOutOfSyncException;
 import com.headstrongpro.desktop.core.exception.ModelSyncException;
 import com.headstrongpro.desktop.model.PaymentRate;
 import com.headstrongpro.desktop.model.Subscription;
 import com.headstrongpro.desktop.model.entity.Company;
+import com.headstrongpro.desktop.modelCollections.util.ActionType;
 import com.headstrongpro.desktop.modelCollections.util.IDataAccessObject;
+import com.headstrongpro.desktop.modelCollections.util.Synchronizable;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -14,7 +17,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DBSubscriptions implements IDataAccessObject<Subscription> {
+public class DBSubscriptions extends Synchronizable implements IDataAccessObject<Subscription> {
 
     private DBConnect dbConnect;
 
@@ -93,6 +96,7 @@ public class DBSubscriptions implements IDataAccessObject<Subscription> {
     public Subscription create(Subscription object) throws ModelSyncException {
         try {
             dbConnect = new DBConnect();
+            //language=TSQL
             String createCompanyQuery = "INSERT INTO subscriptions (company_id, is_active, start_date, end_date, no_of_users, rate_id) VALUES (?, ?, ?, ?, ?, ?);";
             PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(createCompanyQuery, PreparedStatement.RETURN_GENERATED_KEYS);
             preparedStatement.setInt(1, object.getCompany().getId());
@@ -105,6 +109,7 @@ public class DBSubscriptions implements IDataAccessObject<Subscription> {
             try (ResultSet generatedKeys = preparedStatement.getGeneratedKeys()) {
                 if (generatedKeys.next()) {
                     object.setId(generatedKeys.getInt(1));
+                    logChange("subscriptions", object.getId(), ActionType.CREATE);
                 } else {
                     throw new ModelSyncException("Creating subscription failed. No ID retrieved!");
                 }
@@ -116,35 +121,45 @@ public class DBSubscriptions implements IDataAccessObject<Subscription> {
     }
 
     @Override
-    public void update(Subscription object) throws ModelSyncException {
-        try {
-            dbConnect = new DBConnect();
-            //language=TSQL
-            String query = "UPDATE subscriptions SET company_id=?,is_active=?,start_date=?,end_date=?,no_of_users=?,rate_id=? WHERE id=?;";
-            PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
-            preparedStatement.setInt(1, object.getCompany().getId());
-            preparedStatement.setBoolean(2, object.isActive());
-            preparedStatement.setDate(3, object.getStartDate());
-            preparedStatement.setDate(4, object.getEndDate());
-            preparedStatement.setInt(5, object.getNoOfUsers());
-            preparedStatement.setInt(6, object.getRate().getId());
-            dbConnect.uploadSafe(preparedStatement);
-        } catch (ConnectionException | SQLException e) {
-            throw new ModelSyncException("WARNING! Could not update resource of ID: " + object.getId() + " !", e);
+    public void update(Subscription object) throws ModelSyncException, DatabaseOutOfSyncException {
+        if (verifyIntegrity(object.getId())){
+            try {
+                dbConnect = new DBConnect();
+                //language=TSQL
+                String query = "UPDATE subscriptions SET company_id=?,is_active=?,start_date=?,end_date=?,no_of_users=?,rate_id=? WHERE id=?;";
+                PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
+                preparedStatement.setInt(1, object.getCompany().getId());
+                preparedStatement.setBoolean(2, object.isActive());
+                preparedStatement.setDate(3, object.getStartDate());
+                preparedStatement.setDate(4, object.getEndDate());
+                preparedStatement.setInt(5, object.getNoOfUsers());
+                preparedStatement.setInt(6, object.getRate().getId());
+                dbConnect.uploadSafe(preparedStatement);
+                logChange("subscriptions", object.getId(), ActionType.UPDATE);
+            } catch (ConnectionException | SQLException e) {
+                throw new ModelSyncException("WARNING! Could not update resource of ID: " + object.getId() + " !", e);
+            }
+        } else {
+            throw new DatabaseOutOfSyncException();
         }
     }
 
     @Override
-    public void delete(Subscription object) throws ModelSyncException {
-        try {
-            dbConnect = new DBConnect();
-            //language=TSQL
-            String query = "DELETE FROM subscriptions WHERE id=?;";
-            PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
-            preparedStatement.setInt(1, object.getId());
-            preparedStatement.execute();
-        } catch (ConnectionException | SQLException e) {
-            throw new ModelSyncException("WARNING! Could not update subscription of ID: " + object.getId() + " !", e);
+    public void delete(Subscription object) throws ModelSyncException, DatabaseOutOfSyncException {
+        if (verifyIntegrity(object.getId())){
+            try {
+                dbConnect = new DBConnect();
+                //language=TSQL
+                String query = "DELETE FROM subscriptions WHERE id=?;";
+                PreparedStatement preparedStatement = dbConnect.getConnection().prepareStatement(query);
+                preparedStatement.setInt(1, object.getId());
+                preparedStatement.execute();
+                logChange("subscriptions", object.getId(), ActionType.DELETE);
+            } catch (ConnectionException | SQLException e) {
+                throw new ModelSyncException("WARNING! Could not update subscription of ID: " + object.getId() + " !", e);
+            }
+        } else {
+            throw new DatabaseOutOfSyncException();
         }
     }
 
@@ -172,5 +187,10 @@ public class DBSubscriptions implements IDataAccessObject<Subscription> {
             throw new ModelSyncException("Could not load subscriptions.", e);
         }
         return subscriptions;
+    }
+
+    @Override
+    protected boolean verifyIntegrity(int itemID) throws ModelSyncException {
+        return true; //TODO: to be implemented
     }
 }
